"Allow inbound ports" changes
The lab steps instruct students to create a VM and set Public inbound ports to None on the Basics tab, then set Public inbound ports to HTTP and RDP on the Networking tab of the Create a virtual machine blade. But it seems that something has changed in the process, because the "Select inbound ports" drop-down list is non-responsive using this technique. The changes I propose are to allow HTTP and RDP from the Basics tab instead.